BookFinder/backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
Authlib==1.6.5
certifi==2025.11.12
cffi==2.0.0
click==8.3.1
cryptography==46.0.3
dotenv==0.9.9
ecdsa==0.19.1
fastapi==0.121.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
itsdangerous==2.2.0
pyasn1==0.6.1
pycparser==2.23
pydantic==2.12.4
pydantic_core==2.41.5
python-dotenv==1.2.1
python-jose==3.5.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
starlette==0.49.3
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.38.0
requests==2.32.5
pandas==2.3.3
psycopg2-binary==2.9.11
SQLAlchemy==2.0.36
loguru==0.7.2
# DS service dependencies
openai>=1.0.0,<2.0.0
sentence-transformers==3.1.1
numpy>=1.24.0,<2.0.0
faiss-cpu==1.8.0

